refactor(driver): extract verify + filtering methods to driver/verify.rs (refactor 2.1i)
Fourth impl-block extraction of Phase 2.1 and the biggest by LOC.
Moves the eight methods that own the post-derivation verify path
and the §4f deferred filtering machinery.

- `crates/based-rollup/src/driver/verify.rs` (new, 820 lines):
  - `classify_and_apply_verification` — state machine behind
    `verify_local_block_matches_l1`. Owns invariants #9 + #10 closure.
  - `verify_local_block_matches_l1` — thin `Result<()>` wrapper.
  - `apply_deferred_filtering` — top-level router between rebuild
    and filter paths.
  - `apply_generic_filtering` — raw-byte prefix filter.
  - `apply_generic_filtering_via_rebuild` — rebuild from entries
    with `max_trigger_count`; nonce save/restore on fallback.
  - `compute_state_root_with_entries` — per-prefix state root
    computation via an isolated EVM config clone.
  - `trial_execute_for_receipts` — trial execution returning
    per-tx receipts for generic trigger identification.
  - `compute_intermediate_roots` — generic N+1 root chain builder
    for unified D+W state-delta construction.
  - All methods are `pub(super) fn` so the rest of the driver tree
    can still call them via `self.…`.
  - Full `impl<P, Pool> Driver<P, Pool> where …` wrapper with the
    same bounds as mod.rs.

- `crates/based-rollup/src/driver/mod.rs` (-750 lines → 3384 total):
  - `mod verify;` declaration.
  - Eight method bodies removed from the main `impl Driver` block.
  - Dead imports cleaned: `trace` (used only in the thin wrapper
    that moved out), `VerificationDecision` (used only in the
    classify method that moved out).

Cumulative progress:
- mod.rs: 5509 → 3384 lines (-2125 / -38.6%) — **past a third**.
- 8 sibling modules totaling 2442 lines.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Next: 2.1j — protocol_txs.rs (extracts `build_builder_protocol_txs`
and `build_derived_block`, ~400 lines). After that: flush.rs
(~900 lines) and step_builder.rs (~800 lines) are the final two
extractions before mod.rs is reduced to core + `new` + `run` + `step`.
